Moved all port specific functions to datetime_init. Add datetime_timesync function to be implemented in every port. Tested in Linux, implemented but not tested on Win32. Fixed UTC_Offset from the difference between localtime and UTC to difference between local solar time and UTC
