+++ apps/server/Makefile
@@ -5,7 +5,6 @@ TARGET = bacserv
 # BACnet objects that are used with this app
 BACNET_OBJECT_DIR = $(BACNET_SRC_DIR)/bacnet/basic/object
 SRC = main.c \
-	datetime-init.c \
 	$(BACNET_OBJECT_DIR)/device.c \
 	$(BACNET_OBJECT_DIR)/ai.c \
 	$(BACNET_OBJECT_DIR)/ao.c \
